refactor(protobuf): create protobuf crate
