+++ crates/papyrus_network/Cargo.toml
@@ -35,6 +35,7 @@ metrics.workspace = true
 replace_with.workspace = true
 papyrus_common = { path = "../papyrus_common", version = "0.4.0-dev.2" }
 papyrus_config = { path = "../papyrus_config", version = "0.4.0-dev.2" }
+papyrus_protobuf = { path = "../papyrus_protobuf", version = "0.4.0-dev.2" }
 papyrus_storage = { path = "../papyrus_storage", version = "0.4.0-dev.2" }
 prost.workspace = true
 prost-types.workspace = true
