+++ crates/papyrus_network/src/network_manager/mod.rs
@@ -13,6 +13,7 @@ use libp2p::swarm::SwarmEvent;
 use libp2p::{PeerId, Swarm};
 use metrics::gauge;
 use papyrus_common::metrics as papyrus_metrics;
+use papyrus_protobuf::protobuf;
 use papyrus_storage::StorageReader;
 use prost::Message;
 use streamed_bytes::Bytes;
@@ -24,7 +25,6 @@ use crate::broadcast::Topic;
 use crate::converters::{Router, RouterError};
 use crate::db_executor::{self, DBExecutor, DBExecutorTrait, Data, QueryId};
 use crate::mixed_behaviour::{self, BridgedBehaviour};
-use crate::protobuf_messages::protobuf;
 use crate::streamed_bytes::{self, InboundSessionId, OutboundSessionId, SessionId};
 use crate::utils::StreamHashMap;
 use crate::{broadcast, DataType, NetworkConfig, Protocol, Query, ResponseReceivers};
