+++ crates/papyrus_node/Cargo.toml
@@ -34,6 +34,7 @@ papyrus_common = { path = "../papyrus_common", version = "0.4.0-dev.2" }
 papyrus_monitoring_gateway = { path = "../papyrus_monitoring_gateway", version = "0.4.0-dev.2" }
 papyrus_network = { path = "../papyrus_network", version = "0.4.0-dev.2" }
 papyrus_p2p_sync = { path = "../papyrus_p2p_sync", version = "0.4.0-dev.2" }
+papyrus_protobuf = { path = "../papyrus_protobuf", version = "0.4.0-dev.2" }
 papyrus_rpc = { path = "../papyrus_rpc", version = "0.4.0-dev.2", optional = true }
 papyrus_storage = { path = "../papyrus_storage", version = "0.4.0-dev.2" }
 papyrus_sync = { path = "../papyrus_sync", version = "0.4.0-dev.2" }
