+++ crates/papyrus_p2p_sync/Cargo.toml
@@ -14,6 +14,7 @@ papyrus_common = { path = "../papyrus_common", version = "0.4.0-dev.2" }
 papyrus_config = { path = "../papyrus_config", version = "0.4.0-dev.2" }
 papyrus_network = { path = "../papyrus_network", version = "0.4.0-dev.2" }
 papyrus_proc_macros = { path = "../papyrus_proc_macros", version = "0.4.0-dev.2" }
+papyrus_protobuf = { path = "../papyrus_protobuf", version = "0.4.0-dev.2" }
 papyrus_storage = { path = "../papyrus_storage", version = "0.4.0-dev.2" }
 serde.workspace = true
 starknet_api.workspace = true
