+++ crates/papyrus_p2p_sync/src/lib.rs
@@ -15,7 +15,8 @@ use futures::channel::mpsc::{SendError, Sender};
 use papyrus_config::converters::deserialize_seconds_to_duration;
 use papyrus_config::dumping::{ser_optional_param, ser_param, SerializeConfig};
 use papyrus_config::{ParamPath, ParamPrivacyInput, SerializedParam};
-use papyrus_network::{DataType, Query, ResponseReceivers};
+use papyrus_network::{DataType, ResponseReceivers};
+use papyrus_protobuf::sync::Query;
 use papyrus_storage::{StorageError, StorageReader, StorageWriter};
 use serde::{Deserialize, Serialize};
 use starknet_api::block::{BlockNumber, BlockSignature};
