+++ crates/papyrus_p2p_sync/src/state_diff_test.rs
@@ -4,7 +4,8 @@ use assert_matches::assert_matches;
 use futures::future::ready;
 use futures::{FutureExt, SinkExt, StreamExt};
 use indexmap::{indexmap, IndexMap};
-use papyrus_network::{BlockHashOrNumber, DataType, Direction, Query, SignedBlockHeader};
+use papyrus_network::DataType;
+use papyrus_protobuf::sync::{BlockHashOrNumber, Direction, Query, SignedBlockHeader};
 use papyrus_storage::state::StateStorageReader;
 use rand::RngCore;
 use starknet_api::block::{BlockHeader, BlockNumber};
