+++ crates/papyrus_p2p_sync/src/stream_factory.rs
@@ -7,7 +7,8 @@ use futures::channel::mpsc::Sender;
 use futures::future::BoxFuture;
 use futures::stream::BoxStream;
 use futures::{SinkExt, Stream, StreamExt};
-use papyrus_network::{BlockHashOrNumber, DataType, Direction, Query};
+use papyrus_network::DataType;
+use papyrus_protobuf::sync::{BlockHashOrNumber, Direction, Query};
 use papyrus_storage::header::HeaderStorageReader;
 use papyrus_storage::{StorageError, StorageReader, StorageWriter};
 use starknet_api::block::BlockNumber;
