+++ crates/papyrus_p2p_sync/src/test_utils.rs
@@ -3,7 +3,8 @@ use std::time::Duration;
 use futures::channel::mpsc::{Receiver, Sender};
 use futures::StreamExt;
 use lazy_static::lazy_static;
-use papyrus_network::{DataType, Query, ResponseReceivers, SignedBlockHeader};
+use papyrus_network::{DataType, ResponseReceivers};
+use papyrus_protobuf::sync::{Query, SignedBlockHeader};
 use papyrus_storage::test_utils::get_test_storage;
 use papyrus_storage::StorageReader;
 use starknet_api::block::{BlockHash, BlockSignature};
