+++ crates/papyrus_protobuf/src/converters/header.rs
@@ -21,8 +21,8 @@ use starknet_api::crypto::Signature;
 
 use super::common::{enum_int_to_l1_data_availability_mode, l1_data_availability_mode_to_enum_int};
 use super::ProtobufConversionError;
-use crate::protobuf_messages::protobuf::{self};
-use crate::{Query, SignedBlockHeader};
+use crate::protobuf;
+use crate::sync::{Query, SignedBlockHeader};
 
 impl TryFrom<protobuf::BlockHeadersResponse> for Option<SignedBlockHeader> {
     type Error = ProtobufConversionError;
@@ -289,10 +289,12 @@ impl From<Option<SignedBlockHeader>> for protobuf::BlockHeadersResponse {
 impl TryFrom<protobuf::BlockHeadersRequest> for Query {
     type Error = ProtobufConversionError;
     fn try_from(value: protobuf::BlockHeadersRequest) -> Result<Self, Self::Error> {
-        let value = value.iteration.ok_or(ProtobufConversionError::MissingField {
-            field_description: "BlockHeadersRequest::iteration",
-        })?;
-        value.try_into()
+        value
+            .iteration
+            .ok_or(ProtobufConversionError::MissingField {
+                field_description: "BlockHeadersRequest::iteration",
+            })?
+            .try_into()
     }
 }
