+++ crates/papyrus_protobuf/src/converters/mod.rs
@@ -1,5 +1,7 @@
-mod common;
+// TODO(shahak): Internalize this once network doesn't depend on protobuf.
+pub mod common;
 mod header;
+// TODO(shahak): Internalize this once network doesn't depend on protobuf.
 pub mod state_diff;
 mod transaction;
