+++ crates/papyrus_protobuf/src/converters/state_diff.rs
@@ -4,10 +4,11 @@ use starknet_api::data_availability::DataAvailabilityMode;
 use starknet_api::hash::StarkFelt;
 use starknet_api::state::{StorageKey, ThinStateDiff};
 
-use super::common::volition_domain_to_enum_int;
 use super::ProtobufConversionError;
-use crate::protobuf_messages::protobuf;
-use crate::Query;
+use crate::protobuf;
+use crate::sync::Query;
+
+pub const DOMAIN: DataAvailabilityMode = DataAvailabilityMode::L1;
 
 impl TryFrom<protobuf::StateDiffsResponse> for Option<ThinStateDiff> {
     type Error = ProtobufConversionError;
@@ -134,10 +135,12 @@ impl TryFrom<protobuf::ContractStoredValue> for (StorageKey, StarkFelt) {
 impl TryFrom<protobuf::StateDiffsRequest> for Query {
     type Error = ProtobufConversionError;
     fn try_from(value: protobuf::StateDiffsRequest) -> Result<Self, Self::Error> {
-        let value = value.iteration.ok_or(ProtobufConversionError::MissingField {
-            field_description: "StateDiffsRequest::iteration",
-        })?;
-        value.try_into()
+        value
+            .iteration
+            .ok_or(ProtobufConversionError::MissingField {
+                field_description: "StateDiffsRequest::iteration",
+            })?
+            .try_into()
     }
 }
